+++ Services/FirstRunGuideLaunchMode.cs
@@ -1,0 +1,7 @@
+namespace UrbanPlanToolbox.Services;
+
+public enum FirstRunGuideLaunchMode
+{
+    Automatic,
+    Manual
+}